[nrf fromtree] samples: boards: nordic: coresight_stm: Fix nrfutil trace command
Latest version of nrfutil trace doesn't support setting
uart baudrate.
Remove --baudrate {baudrate_val} from command that executes nrfutil.

Also, combine pytest script for:
- samples.boards.nordic.coresight_stm,
- tests.boards.nrf.coresight_stm,
to minimize code duplication (fix compliance check R0801).

(cherry picked from commit 153f5b63820ad2ec96c31ae0f868323a2c585998)
