+++ tests/boards/nrf/coresight_stm/testcase.yaml
@@ -16,7 +16,7 @@ tests:
     harness_config:
       pytest_dut_scope: session
       pytest_root:
-        - "pytest/test_stm.py::test_STM_dictionary_mode"
+        - "$ZEPHYR_BASE/samples/boards/nordic/coresight_stm/pytest/test_stm.py::test_STM_dictionary_mode"
     required_snippets:
       - nordic-log-stm-dict
     extra_args:
@@ -29,7 +29,7 @@ tests:
     harness_config:
       pytest_dut_scope: session
       pytest_root:
-        - "pytest/test_stm.py::test_STM_decoded"
+        - "$ZEPHYR_BASE/samples/boards/nordic/coresight_stm/pytest/test_stm.py::test_STM_decoded"
     required_snippets:
       - nordic-log-stm
     extra_args:
